There's one more cool thing I want to show you with integration tests. It's kind of a weird, unnatural mixture of unit tests and integration tests. I love it! I call it partial mocking.
Right now, we're fetching the EnclosureBuilderService from the container. But there's another option. Instead, create it manually - new EnclosureBuilderService() - and pass in the dependencies manually. For example, pass ->getEntityManager() as the first argument.

Why would we do this? It seems like more work! Because... if it's useful, we could mock certain dependencies. Yep! Instead of fetching the DinosaurFactory from the container, mock it.

Pass dinoFactory variable this as the second argument. This isn't particularly useful in this situation. But sometimes, being able to mock - and control - one or two dependencies in an integration test is really awesome!

Ok, try the test!

./vendor/bin/phpunit --filter testItBuildsEnclosureWithTheDefaultSpecification
Oh man! It fails! Weird! There is only 1 Dinosaur... but there should be 3! What's going on? This is subtle: PhpUnit is smart enough to take this one dinosaur object and return it each time growFromSpecification() is called. But to Doctrine, it looks like we're asking it to save the same one Dinosaur object: not three separate dinosaurs. The result would be a less than thrilling theme park.