The fix of the previous commit problem is to change this to willReturnCallback() and pass it a function. This will be called each time growFromSpecification() is called. And since it is passed a  argument, the callback also receives this. It's the best way to return different values based on the arguments. We don't need that in this case. We'll just say return new Dinosaur(). And that's it! Try the tests again. ./vendor/bin/phpunit --filter testItBuildsEnclosureWithTheDefaultSpecification We got it! So integration tests are one of my favorite, favorite, favorite things because they're very pragmatic. When I think about testing a class, here's the logic I follow:
Does this class scare me? If the logic is simple, I don't test it. I will still make sure the new feature works through manual testing. And often, that's enough.

Can I unit test this class? If I mock all of the dependencies, is the scary logic still testable? For DinosaurFactory, the answer was yes: we mocked the DinosaurLengthDeterminator, but we were still able to test that the specification string is parsed correctly to create carnivorous and non-carnivorous dinosaurs.

If the class cannot be unit tested, then I write an integration test. This is pretty common in my apps: each individual 'unit' is often pretty simple. But when you integrate them all together, that's scary enough to need a test.

Ok, on to functional testing!