Writing the Enclosure Test
In the test, we need to think about how this might look. Let's add {dollar_sign}table = {dollar_sign}crawler->filter() and then a CSS selector: .table-enclosures. I'm saying that, when we build this page, we should create an element with this class. The {dollar_sign}crawler object is a bit similar to the jQuery function in JavaScript: by using its filter() method, it's really good at finding elements via CSS. The table variable is itself another Crawler object, which represents the table element.
Now, we can assertCount() that 3 is equal to {dollar_sign}table->filter('tbody tr'). In other words, inside the table, we expect there to be 3 rows. Why 3? Well... I just made that up! Just like with integration tests, we're going to need to take control of the database so that we know exactly what's inside. More on that soon!

Try the test now:

./vendor/bin/phpunit tests/AppBundle/Controller/DefaultControllerTest.php
Yay! It fails! That means we are ready to code!