Open up Enclosure. Near the top, at getId(): it should return a nullable int. And at the bottom, create a public function getDinosaurCount() that will return an int.
So... does the page work? It should! Try the tests!

./vendor/bin/phpunit tests/AppBundle/Controller/DefaultControllerTest.php
Ah! Failure because the actual size one does not match expected size 3. Check out the homepage in the browser: we have 7 enclosures! This is the same problem we had with integration test: we're not taking control of the data in the database. So... we really have no idea how many enclosures will be in the list! Our main database has 7... and apparently our test database has only one.

Let's fix this!