Despite all our precautions, we still sometimes have enclosures with no security. Yea... a lot of people are getting eaten, a lot of lawsuits - very expensive. To help with this, I want to add an 'Alarm' button on the homepage next to any enclosures that do not have active security.
What are Fixture References?
Here's the deal: if you look in the fixtures, you can see that the first two Enclosures do not have any security. You can also see that we're using some sort of 'reference' system. This allows us to store a specific object in memory so that we can re-use it somewhere else. For example, in LoadSecurityData, we get the herbivorous-enclosure object out and add security! We're safe from those wild veggie eating dinos!

It does the same for carnivorous-enclosure... but then adds two Security objects that are both inactive. Doh! Yep, this means that the carnivorous-enclosure is the only Enclosure that is not secure. In the test, our goal is to assert that, on the homepage, this exact Enclosure has the alarm button.

And we planned ahead for this! Remember, in the template, we added an enclosure-{id} to each tr element. So if we can get the actual id value of the Carnivorous Enclosure, it will be really easy to find its tr element and look for the alarm button. The reference system gives us that power!