Debugging Functional Tests But... what if it didn't pass? Well... the errors wouldn't be very helpful: it would basically just say that 0 is not greater than 0. When things fail, the trick is to go above the failure and dump({dollar_sign}client->getResponse()->getContent()). If you're using Flex, make sure to install the var-dumper package.
Now when you run the test, it will at least print out the HTML body. By the way, with a little bit of clever coding, you can hook into the onNotSuccessfulTest method and have the last response content printed automatically when a test fails. I'll leave that as a challenge for you. But, ask us in the comments if you have any questions.

Ok, there's one more thing I want to talk about with functional tests: filling out and submitting a form.