New feature request! On the homepage, management wants a form where they can choose an enclosure, write a dinosaur spec - like 'Large herbivore' and submit! Behind the scenes, we will create that new Dinosaur and put it into the Enclosure.
Since we're now functional-testing pros, let's get right to the test! Add public function testItGrowsADinosaurFromSpecification(). And as usual, steal some code from earlier and paste it on top. You can start to see how some of this could be refactored to a setUp method.
After creating the client, add {dollar_sign}client->followRedirects(). Normally, when our app redirects, Symfony's Client does not follow the redirect. Sometimes that's useful... but this line makes it behave like a normal browser.

Filling in the Form Fields
To fill out the form fields, first we need to find the form. Do that with {dollar_sign}form = {dollar_sign}crawler->selectButton() and pass this the value of the button that will be on your form. How about 'Grow dinosaur'. Then call ->form().

We now have a Form object. No, not Symfony's normal Form object from its form system. This is from the DomCrawler component and its job is to help us fill out its fields.

So let's think about it: we will need 2 fields: an enclosure select field and a specification text box. To fill in the first, use {dollar_sign}form['enclosure'] - the enclosure part is whatever the name attribute for your field will be. If you're using Symfony forms, usually this will look more like dinosuar[enclosure].

Then, because this will be a select field, use ->select(3), where 3 is the value of the option element you want to select. Do this again for a specification field. Setting this one is easier: ->setValue('large herbivore').

Honestly, I don't love Symfony's API for filling in forms - I like Mink's better. But, it works fine. When the form is ready, submit with {dollar_sign}client->submit({dollar_sign}form). That will submit to the correct URL and send all the data up!

But... now what? What should the user see after submitting the form? Well... we should probably redirect back to the homepage with a nice message explaining what just happened. Use {dollar_sign}this->assertContains() to look for the text 'Grew a large herbivore in enclosure #3' inside {dollar_sign}client->getResponse()->getContent().

Test, done! Copy the method name and just run this test:

./vendor/bin/phpunit --filter testItGrowsADinosaurFromSpecification
Perfect! It fails with

The current node list is empty.

This is a really common error... though it's not the most helpful. It basically means that some element could not be found.